Fix bug in handling empty memop log.

+++ mem.h
@@ -22,8 +22,8 @@ extern "C" {
 #define OBJ_SIZE 8
 #define NOBJS 10
 
-typedef long version_t;
-typedef int16_t objid_t;
+typedef int version_t;
+typedef int8_t objid_t;
 typedef int memop_t;
 typedef int8_t tid_t;
